Separate the menu wrapper out to a separate file
