+++ packages/dropdowns/.size-snapshot.json
@@ -1,20 +1,20 @@
 {
   "dist/index.cjs.js": {
-    "bundled": 76487,
-    "minified": 49163,
-    "gzipped": 10206
+    "bundled": 75937,
+    "minified": 48859,
+    "gzipped": 10086
   },
   "dist/index.esm.js": {
-    "bundled": 72894,
-    "minified": 45683,
-    "gzipped": 10004,
+    "bundled": 72349,
+    "minified": 45384,
+    "gzipped": 9890,
     "treeshaked": {
       "rollup": {
-        "code": 35754,
+        "code": 35495,
         "import_statements": 788
       },
       "webpack": {
-        "code": 39474
+        "code": 39243
       }
     }
   }
